requirements.txt
# Structure
pyyaml>=6.0
pytest>=7.0
pytest-mock>=3.14
pyarrow>=13.0.0,<16.0.0

# Base
pandas>=1.5.3,<2.2
numpy>=1.25
scikit-learn>=1.3
xgboost>=1.7.0
torch>=2.1.0

# Visualisation
matplotlib>=3.7
